requirements.txt
Flask==2.0.2
gunicorn
Werkzeug==2.2.2
openai==0.28.1
python-dotenv==1.0.0